requirements.txt
requests~=2.31.0
aiohttp~=3.8.6
discord~=2.3.2
beautifulsoup4~=4.12.2
python-dotenv~=1.0.0